+++ tests/beaker_tests/routedintf/routedintf-provider-nondefaults.rb
@@ -116,6 +116,7 @@ test_name "TestCase :: #{testheader}" do
         'ipv4_netmask_length'          => '16',
         'ipv4_proxy_arp'               => 'true',
         'ipv4_redirects'               => 'false',
+        'mtu'                          => '1556',
         'shutdown'                     => 'true',
         'switchport_autostate_exclude' => 'false',
         'switchport_mode'              => 'disabled',
@@ -131,13 +132,102 @@ test_name "TestCase :: #{testheader}" do
   step "TestStep :: Check interface instance presence on agent" do
     # Expected exit_code is 0 since this is a vegas shell cmd.
     # Flag is set to false to check for presence of RegExp pattern in stdout.
-    cmd_str = UtilityLib.get_vshell_cmd("show running-config interface eth1/4")
+    cmd_str = UtilityLib.get_vshell_cmd("show running-config interface eth1/4 all")
     on(agent, cmd_str) do
       UtilityLib.search_pattern_in_output(stdout, [/ip address 192.168.1.1\/16/,
         /no switchport/,
         /no ip redirects/,
         /ip proxy-arp/,
+        /mtu 1556/,
         /vrf member test1/],
+        false, self, logger)
+    end
+
+    logger.info("Check interface instance presence on agent :: #{result}")
+  end
+
+  # @step [Step] Requests manifest from the master server to the agent.
+  step "TestStep :: Get resource nondefaults manifest from master" do
+    # Expected exit_code is 0 since this is a bash shell cmd.
+    on(master, RoutedIntfLib.create_routedintf_manifest_subinterface())
+
+    # Expected exit_code is 2 since this is a puppet agent cmd with change.
+    cmd_str = UtilityLib.get_namespace_cmd(agent, UtilityLib::PUPPET_BINPATH +
+      "agent -t", options)
+    on(agent, cmd_str, {:acceptable_exit_codes => [2]})
+
+    logger.info("Get resource nondefaults manifest from master :: #{result}")
+  end
+
+  # @step [Step] Checks cisco_interface resource on agent using resource cmd.
+  step "TestStep :: Check cisco_interface resource presence on agent" do
+    # Expected exit_code is 0 since this is a puppet resource cmd.
+    # Flag is set to false to check for presence of RegExp pattern in stdout.
+    cmd_str = UtilityLib.get_namespace_cmd(agent, UtilityLib::PUPPET_BINPATH +
+      "resource cisco_interface 'ethernet1/4.1'", options)
+    on(agent, cmd_str) do
+      UtilityLib.search_pattern_in_output(stdout,
+        {'ensure'                      => 'present',
+        'encapsulation_dot1q'          => '30'},
+        false, self, logger)
+    end
+
+    logger.info("Check cisco_interface resource presence on agent :: #{result}")
+  end
+
+  # @step [Step] Checks interface instance on agent using switch show cli cmds.
+  step "TestStep :: Check interface instance presence on agent" do
+    # Expected exit_code is 0 since this is a vegas shell cmd.
+    # Flag is set to false to check for presence of RegExp pattern in stdout.
+    cmd_str = UtilityLib.get_vshell_cmd("show running-config interface eth1/4.1")
+    on(agent, cmd_str) do
+      UtilityLib.search_pattern_in_output(stdout, [/encapsulation dot1q 30/],
+        false, self, logger)
+    end
+
+    logger.info("Check interface instance presence on agent :: #{result}")
+  end
+
+  # @step [Step] Requests manifest from the master server to the agent.
+  step "TestStep :: Get resource nondefaults manifest from master" do
+    # Expected exit_code is 0 since this is a bash shell cmd.
+    on(master,
+        RoutedIntfLib.create_routedintf_manifest_switchport_trunk_nondefaults())
+
+    # Expected exit_code is 2 since this is a puppet agent cmd with change.
+    cmd_str = UtilityLib.get_namespace_cmd(agent, UtilityLib::PUPPET_BINPATH +
+      "agent -t", options)
+    on(agent, cmd_str, {:acceptable_exit_codes => [2]})
+
+    logger.info("Get resource nondefaults manifest from master :: #{result}")
+  end
+
+  # @step [Step] Checks cisco_interface resource on agent using resource cmd.
+  step "TestStep :: Check cisco_interface resource presence on agent" do
+    # Expected exit_code is 0 since this is a puppet resource cmd.
+    # Flag is set to false to check for presence of RegExp pattern in stdout.
+    cmd_str = UtilityLib.get_namespace_cmd(agent, UtilityLib::PUPPET_BINPATH +
+      "resource cisco_interface 'ethernet1/4'", options)
+    on(agent, cmd_str) do
+      UtilityLib.search_pattern_in_output(stdout,
+        {'ensure'                       => 'present',
+        'switchport_trunk_allowed_vlan' => '30,40',
+        'switchport_trunk_native_vlan'  => '20'},
+        false, self, logger)
+    end
+
+    logger.info("Check cisco_interface resource presence on agent :: #{result}")
+  end
+
+  # @step [Step] Checks interface instance on agent using switch show cli cmds.
+  step "TestStep :: Check interface instance presence on agent" do
+    # Expected exit_code is 0 since this is a vegas shell cmd.
+    # Flag is set to false to check for presence of RegExp pattern in stdout.
+    cmd_str = UtilityLib.get_vshell_cmd("show running-config interface eth1/4")
+    on(agent, cmd_str) do
+      UtilityLib.search_pattern_in_output(stdout, 
+        [/switchport trunk allowed vlan 30,40/,
+        /switchport trunk native vlan 20/],
         false, self, logger)
     end