+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- Extended cisco_interface with the following attributes:
+  - encapsulation dot1q
+  - mtu
+  - switchport trunk allowed VLANs
+  - switchport trunk native VLAN
+
 ## [1.0.0] - 2015-08-28
 ### Added
 - New facts `cisco_node_utils` and `cisco_nxapi` report the installed version of these gems.